Fix compilation errors

+++ game.h
@@ -8,15 +8,15 @@ extern "C" {
 #include <stdint.h>
 #endif
 
-void fill_next_state(const char* board, char* board_next, ssize_t width, ssize_t height);
+void fill_next_state(const char* board, char* board_next, int width, int height);
 
 void fill_pixels(const char* board,
-                 ssize_t board_width,
-                 ssize_t board_height,
+                 int board_width,
+                 int board_height,
                  unsigned char* pixels,
-                 ssize_t surface_width,
-                 ssize_t surface_height,
-                 ssize_t bytes_per_pixel);
+                 int surface_width,
+                 int surface_height,
+                 int bytes_per_pixel);
 
 #ifdef __cplusplus
 }
